feat: refactor worker scripts out of the report svc
