+++ src/reports/reports.service.ts
@@ -126,38 +126,7 @@ export class ReportsService {
   ): Promise<Record<string, number>> {
     return new Promise((resolve, reject) => {
       const worker = new Worker(
-        `
-        const { parentPort } = require('worker_threads');
-        const { createReadStream } = require('fs');
-        const { createInterface } = require('readline');
-        const path = require('path');
-
-        async function processFiles(files, tmpDir) {
-          const balances = {};
-
-          for (const file of files) {
-            const fileStream = createReadStream(path.join(tmpDir, file));
-            const rl = createInterface({ input: fileStream });
-
-            for await (const line of rl) {
-              const [, account, , debit, credit] = line.split(',');
-              balances[account] = (balances[account] || 0) +
-                parseFloat(String(debit || 0)) - parseFloat(String(credit || 0));
-            }
-          }
-          return balances;
-        }
-
-        parentPort.on('message', async ({ files, tmpDir }) => {
-          try {
-            const result = await processFiles(files, tmpDir);
-            parentPort.postMessage(result);
-          } catch (error) {
-            parentPort.postMessage({ error: error.message });
-          }
-        });
-      `,
-        { eval: true },
+        path.resolve(__dirname, 'workers', 'account.js'),
       );
 
       worker.postMessage({ files, tmpDir });
@@ -179,44 +148,7 @@ export class ReportsService {
   ): Promise<Record<string, number>> {
     return new Promise((resolve, reject) => {
       const worker = new Worker(
-        `
-        const { parentPort } = require('worker_threads');
-        const { createReadStream } = require('fs');
-        const { createInterface } = require('readline');
-        const path = require('path');
-
-        async function processFiles(files, tmpDir) {
-          const cashByYear = {};
-
-          for (const file of files) {
-            const fileStream = createReadStream(path.join(tmpDir, file));
-            const rl = createInterface({ input: fileStream });
-
-            for await (const line of rl) {
-              const [date, account, , debit, credit] = line.split(',');
-              if (account === 'Cash') {
-                const year = new Date(date).getFullYear();
-                if (!cashByYear[year]) {
-                  cashByYear[year] = 0;
-                }
-                cashByYear[year] +=
-                  parseFloat(String(debit || 0)) - parseFloat(String(credit || 0));
-              }
-            }
-          }
-          return cashByYear;
-        }
-
-        parentPort.on('message', async ({ files, tmpDir }) => {
-          try {
-            const result = await processFiles(files, tmpDir);
-            parentPort.postMessage(result);
-          } catch (error) {
-            parentPort.postMessage({ error: error.message });
-          }
-        });
-      `,
-        { eval: true },
+        path.resolve(__dirname, 'workers', 'yearly.js'),
       );
 
       worker.postMessage({ files, tmpDir });
@@ -238,40 +170,7 @@ export class ReportsService {
     categories: Record<string, Record<string, string[]>>,
   ): Promise<Record<string, number>> {
     return new Promise((resolve, reject) => {
-      const worker = new Worker(
-        `
-        const { parentPort } = require('worker_threads');
-        const { createReadStream } = require('fs');
-        const { createInterface } = require('readline');
-        const path = require('path');
-
-        async function processFiles(files, tmpDir, balance) {
-          for (const file of files) {
-            const fileStream = createReadStream(path.join(tmpDir, file));
-            const rl = createInterface({ input: fileStream });
-
-            for await (const line of rl) {
-              const [, account, , debit, credit] = line.split(',');
-              if (balance.hasOwnProperty(account)) {
-                balance[account] +=
-                  parseFloat(String(debit || 0)) - parseFloat(String(credit || 0));
-              }
-            }
-          }
-          return balance;
-        }
-
-        parentPort.on('message', async ({ files, tmpDir, balance }) => {
-          try {
-            const result = await processFiles(files, tmpDir, balance);
-            parentPort.postMessage(result);
-          } catch (error) {
-            parentPort.postMessage({ error: error.message });
-          }
-        });
-      `,
-        { eval: true },
-      );
+      const worker = new Worker(path.resolve(__dirname, 'workers', 'fs.js'));
       const balance = this.createEmptyBalanceSheet(categories);
       worker.postMessage({ files, tmpDir, balance });
       worker.on('message', (result: WorkerResponse) => {
